started moving db-only models into db folder
This is something I need to start working towards, so that Models really are only models encapsulating the model-related business logic, not worrying about the low-level interactions with the database. Should make it easier to refactor too, since all db-related code is in one place and all model-related code is in another. Clearer separation of concerns